requirements.txt
pydantic==2.12.5
pydantic-settings==2.12.0
requests==2.32.5
schedule==1.2.2

# Image processing libraries
Pillow==12.1.0
HEIC2PNG==1.2.0

# AI libraries
pydantic-ai==1.40.0